requirements.txt
httpx>=0.28.1
langchain-community>=0.3.19
langchain-experimental>=0.3.4
langchain-openai>=0.3.8
langgraph>=0.3.5
readabilipy>=0.3.0
python-dotenv>=1.0.1
socksio>=1.0.0
markdownify>=1.1.0
fastapi>=0.110.0
uvicorn>=0.27.1
sse-starlette>=1.6.5
pandas>=2.2.3
numpy>=1.26.0
yfinance>=0.2.54
litellm>=1.63.11
json-repair>=0.7.0
jinja2>=3.1.3
duckduckgo-search>=8.0.0
inquirerpy>=0.3.4
arxiv>=2.2.0
langchain-mcp-adapters